+++ src/main/groovy/br/com/bluesoft/bee/database/reader/PostgresDatabaseReader.groovy
@@ -85,7 +85,7 @@ class PostgresDatabaseReader implements DatabaseReader {
         def tables = fillTables(objectName)
         fillColumns(tables, objectName)
         fillIndexes(tables, objectName, databaseVersion)
-        fillCostraints(tables, objectName)
+        fillCostraints(tables, objectName, databaseVersion)
         fillCostraintsColumns(tables, objectName)
         return tables
     }
@@ -331,7 +331,7 @@ class PostgresDatabaseReader implements DatabaseReader {
         order by tc.table_name, tc.constraint_type, tc.constraint_name
 	'''
 
-    private def fillCostraints(tables, objectName) {
+    private def fillCostraints(tables, objectName, databaseVersion) {
         def rows
         if (objectName) {
             rows = sql.rows(CONSTRAINTS_QUERY_BY_NAME, [objectName])
@@ -353,7 +353,11 @@ class PostgresDatabaseReader implements DatabaseReader {
             constraint.status = status
             if (constraint.type == 'C') {
                 constraint.refTable = null
-                constraint.searchCondition = it.check_clause[2..-3]
+                if (databaseVersion && Float.parseFloat(databaseVersion) >= 17) {
+                    constraint.searchCondition = it.check_clause[1..-2]
+                } else {
+                    constraint.searchCondition = it.check_clause[2..-3]
+                }
             }
 
             table.constraints[constraint.name] = constraint