+++ src/main/groovy/br/com/bluesoft/bee/schema/BeePostgresSchemaCreatorAction.groovy
@@ -1,6 +1,6 @@
 package br.com.bluesoft.bee.schema
 
-import br.com.bluesoft.bee.importer.JsonImporter
+import br.com.bluesoft.bee.importer.BeeImporter
 import br.com.bluesoft.bee.model.Options
 import br.com.bluesoft.bee.model.Schema
 import br.com.bluesoft.bee.runner.ActionRunner
@@ -88,7 +88,7 @@ public class BeePostgresSchemaCreatorAction implements ActionRunner {
 
     private def getImporter() {
         if (importer == null) {
-            return new JsonImporter(options.dataDir.canonicalPath)
+            return new BeeImporter(options.dataDir.canonicalPath)
         }
         return importer
     }